add tailwind plugin recommendation

+++ generators/vscode/index.js
@@ -1,7 +1,27 @@
 const Generator = require("yeoman-generator");
 
 module.exports = class extends Generator {
-  vscode() {
-    this.fs.copy(this.templatePath(".vscode"), this.destinationPath(".vscode"));
+  async prompting() {
+    this.answer = await this.prompt([
+      {
+        default: true,
+        message: "Does your project use Tailwindcss?",
+        name: "tailwind",
+        type: "confirm"
+      }
+    ])
+  }
+
+  configuring() {
+    if (this.answer.tailwind) {
+      const extensions = this.fs.readJSON(this.templatePath(".extensions.json"))
+
+      extensions.recommendations.push("bradlc.vscode-tailwindcss")
+
+      this.fs.writeJSON(this.destinationPath(".vscode/.extensions.json"), extensions)
+      this.fs.copy(this.templatePath(".vscode/extensions"), this.destinationPath(".vscode/extensions"));
+    } else {
+      this.fs.copy(this.templatePath(".vscode"), this.destinationPath(".vscode"));
+    }
   }
 };
